// src/interface.ts
export interface IApiSignset {
  signsetid: string;
  pkgid: string;
  signsetname: string;
  status: string;
  signerid: string;
  name: string;
  email: string;
  role: string;
  tier: number;
  authcode: string;
  cryptid: string;
  authcoderequired: string;
  verifyrequired: string;
  idprovider: string;
  rolehidden: string;
  required: string;
  delegate: boolean;
  createdby: string;
  createdon: string;
}
// ...
export interface IApiSigboxTemplate {
  //for templates
  left: string;
  top: string;
  width: string;
  height: string;
  esigntype: string;
  boxid: string;
  pagenumber: string;
  font: string;
  fontsize: string;
  lineheight: string;
  fontcolor: string;
  signsetid: string;
  fieldname: string;
  fieldvalue: string;
  fieldlabel: string;
  defaultvalue: string;
  depfield: string;
  depfieldvalue: string;
  depoperator: string;
  fieldrequired: string;
  checkedvalue: string;
  uncheckedvalue: string;
  role: string;
  signerid: number;
  page: number;
  xp: number;
  yp: number;
  hp: number;
  wp: number;
  type: number;
  text: string;
  checked: boolean;
  required: boolean;
}

export interface IParticipants {
  display_name: string;
  fullname: string;
  signing_order: string;
  email: string;
  text_number?: string;
  notification_type: string;
  auth_code: string;
  id_check: string;
  allow_delegation: string;
  key: string;
  color?: string;
  account?: string;
  isSaved?: boolean;
}
// ...
export interface ISigningDoc {
  display_name: string;
  is_template?: boolean;
  sharingids: any[];
  sigboxes: IApiSigboxTemplate[];
  signsets: IApiSignset[];
  targettable: string;
  triggerdefs: any[];
  uniqnm: string;
  tempId?: string;
}

export interface IReviewDoc {
  display_name: string;
  file_name: string;
  is_template?: boolean;
  sharingids: any[];
  sigboxes: IApiSigboxTemplate[];
  signsets: IApiSignset[];
  targettable: string;
  triggerdefs: any[];
  uniqnm: string;
}

export interface IRequestDoc {
  name: string;
  message: string;
}

export interface IElement {
  y: number;
  x: number;
  width: number;
  height: number;
  name: string;
  signer: string;
  options: IOptions[];
  id: string;
  type: string;
  required?: boolean;
  backgroundColor?: string;
  checkedvalue?: string;
  uncheckedvalue?: string;
  defaultvalue?: string;
}

export interface IOptions {
  label: string;
  value: string;
}